+++ compliance-api/.vscode/launch.json
@@ -1,0 +1,16 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Python: Flask",
+      "type": "debugpy",
+      "request": "launch",
+      "module": "flask",
+      "env": { "FLASK_APP": "wsgi.py", "FLASK_ENV": "development" },
+      "args": ["run", "--no-debugger", "--no-reload", "--port", "3200"],
+      "jinja": true,
+      "justMyCode": true,
+      "gevent": true
+    }
+  ]
+}
