+++ compliance-web/src/components/App/RequirementsGrid/RequirementsGridPagination.tsx
@@ -1,0 +1,50 @@
+import { InspectionRequirementGrid } from "@/models/InspectionRequirementGrid";
+import { ChevronLeftRounded, ChevronRightRounded } from "@mui/icons-material";
+import { Box, Typography, IconButton } from "@mui/material";
+import { MRT_TableInstance } from "material-react-table";
+
+interface RequirementsGridPaginationProps {
+  table: MRT_TableInstance<InspectionRequirementGrid>;
+  totalCount: number;
+}
+
+export default function RequirementsGridPagination({
+  table,
+  totalCount,
+}: RequirementsGridPaginationProps) {
+  const currentPage = table.getState().pagination.pageIndex;
+  const pageSize = table.getState().pagination.pageSize;
+  const startItem = currentPage * pageSize + 1;
+  const endItem = Math.min((currentPage + 1) * pageSize, totalCount);
+
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        alignItems: "center",
+        gap: 0.5,
+        justifyContent: "center",
+      }}
+    >
+      <Typography variant="body1">
+        {startItem} to {endItem} of {totalCount}
+      </Typography>
+      <IconButton
+        aria-label="page_back"
+        size="small"
+        onClick={() => table.previousPage()}
+        disabled={!table.getCanPreviousPage()}
+      >
+        <ChevronLeftRounded fontSize="small" />
+      </IconButton>
+      <IconButton
+        aria-label="page_forward"
+        size="small"
+        onClick={() => table.nextPage()}
+        disabled={!table.getCanNextPage()}
+      >
+        <ChevronRightRounded fontSize="small" />
+      </IconButton>
+    </Box>
+  );
+}
