+++ compliance-web/src/routes/_authenticated/ce-database/requirements/index.tsx
@@ -1,12 +1,10 @@
 import MasterDataTable from "@/components/Shared/MasterDataTable/MasterDataTable";
-import PageLink from "@/components/Shared/PageLink";
 import { useRequirementSourcesData } from "@/hooks/useComplaints";
 import {
   useComplianceFindingsData,
   useEnforcementActionsData,
 } from "@/hooks/useInspectionRequirements";
 import {
-  useInspectionRequirementExport,
   useInspectionRequirementsGrid,
 } from "@/hooks/useInspectionRequirementsGrid";
 import { useTopicsData } from "@/hooks/useTopics";
@@ -14,28 +12,20 @@ import {
   InspectionRequirementGrid,
   InspectionRequirementGridQueryParams,
 } from "@/models/InspectionRequirementGrid";
-import { APPROVAL_STATUS, APPROVAL_STATUS_TEXT } from "@/utils/constants";
-import dateUtils from "@/utils/dateUtils";
-import {
-  ChevronLeftRounded,
-  ChevronRightRounded,
-  FileDownloadRounded,
-} from "@mui/icons-material";
-import {
-  Box,
-  Chip,
-  CircularProgress,
-  Typography,
-  Button,
-  IconButton,
-} from "@mui/material";
+import { APPROVAL_STATUS_TEXT } from "@/utils/constants";
+import { Box, CircularProgress, Typography } from "@mui/material";
 import { createFileRoute } from "@tanstack/react-router";
 import { BCDesignTokens } from "epic.theme";
-import { MRT_ColumnDef, MRT_TableState } from "material-react-table";
+import { MRT_TableState } from "material-react-table";
 import { useMemo, useState, useCallback, useEffect, useRef } from "react";
 import RequirementsExternalFilters from "@/components/App/RequirementsGrid/RequirementsExternalFilters";
 import ShowOnlyMyRequirementsSwitch from "@/components/App/RequirementsGrid/ShowOnlyMyRequirementsSwitch";
-import { downloadFile } from "@/utils/appUtils";
+import RequirementsGridPagination from "@/components/App/RequirementsGrid/RequirementsGridPagination";
+import RequirementsGridExport from "@/components/App/RequirementsGrid/RequirementsGridExport";
+import {
+  useConvertFiltersToQueryParams,
+  useRequirementsGridColumns,
+} from "@/components/App/RequirementsGrid/RequirementsGridUtils";
 import { useStaffUsersData } from "@/hooks/useStaff";
 import { cachedFiltersStore } from "@/store/cachedFiltersStore";
 import { AppConfig } from "@/utils/config";
@@ -170,106 +160,9 @@ function Requirements() {
     }
   }, [columnFilters, externalFilters, showOnlyMyRequirements, globalFilter]);
 
-  // Convert column filters to API query parameters
-  const convertFiltersToQueryParams = useCallback(
-    (filters: MRT_TableState<InspectionRequirementGrid>["columnFilters"]) => {
-      const params: Partial<InspectionRequirementGridQueryParams> = {};
-
-      filters.forEach((filter) => {
-        switch (filter.id) {
-          case "topic":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.tpc_ids = filter.value.join(",");
-            }
-            break;
-          case "summary":
-            if (typeof filter.value === "string" && filter.value.trim()) {
-              params.summary = filter.value.trim();
-            }
-            break;
-          case "compliance_finding":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.cmd_fnd_ids = filter.value.join(",");
-            }
-            break;
-          case "enforcement_action":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.enf_actn_ids = filter.value.join(",");
-            }
-            break;
-          case "approval_status":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.apprv_sts = filter.value.join(",");
-            }
-            break;
-          case "reviewer":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.reviewer_ids = filter.value.join(",");
-            }
-            break;
-          case "requirement_number":
-            if (typeof filter.value === "string" && filter.value.trim()) {
-              params.req_src_num = filter.value.trim();
-            }
-            break;
-          case "requirement_source":
-            if (Array.isArray(filter.value) && filter.value.length > 0) {
-              params.req_src_ids = filter.value.join(",");
-            }
-            break;
-          case "ir_number":
-            if (typeof filter.value === "string" && filter.value.trim()) {
-              params.ir_no = filter.value.trim();
-            }
-            break;
-          case "date_issued":
-            if (typeof filter.value === "string" && filter.value.trim()) {
-              // Convert the formatted date back to ISO format for the API
-              const dateValue = filter.value.trim();
-              if (dateValue) {
-                // Assuming the date is in the format used by dateUtils.formatDate
-                // You may need to adjust this based on your API expectations
-                params.date_issued = dateValue;
-              }
-            }
-            break;
-          // requirement_number is not in the query params interface, so skip it
-        }
-      });
-
-      // Add external filters
-      Object.entries(externalFilters).forEach(([key, value]) => {
-        if (value && (Array.isArray(value) ? value.length > 0 : value !== "")) {
-          switch (key) {
-            case "primary_officer_id":
-              params.prm_offc_ids = Array.isArray(value)
-                ? value.join(",")
-                : value;
-              break;
-            case "reviewer_ids":
-              params.reviewer_ids = Array.isArray(value)
-                ? value.join(",")
-                : value;
-              break;
-            case "approval_status":
-              params.apprv_sts = Array.isArray(value) ? value.join(",") : value;
-              break;
-            case "inspection_status":
-              params.insp_sts = Array.isArray(value) ? value.join(",") : value;
-              break;
-            case "project_id":
-              params.project_ids = Array.isArray(value)
-                ? value.join(",")
-                : value;
-              break;
-          }
-        }
-      });
-
-      return params;
-    },
-    [externalFilters]
-  );
+  // Use the extracted utility function
+  const convertFiltersToQueryParams =
+    useConvertFiltersToQueryParams(externalFilters);
 
   const queryParams: InspectionRequirementGridQueryParams = useMemo(
     () => ({
@@ -289,15 +182,6 @@ function Requirements() {
 
   const { data, isLoading } = useInspectionRequirementsGrid(queryParams);
   const requirementsList = useMemo(() => data?.items ?? [], [data]);
-
-  const { mutate: downloadRequirementExport } = useInspectionRequirementExport(
-    (data) => {
-      downloadFile(
-        data,
-        `requirements-${dateUtils.formatDate(new Date().toISOString(), "YYYY-MM-DD-HH-mm-ss")}.xlsx`
-      );
-    }
-  );
 
   const handlePaginationChange = useCallback(
     (
@@ -378,7 +262,7 @@ function Requirements() {
     setPagination((prev) => ({ ...prev, pageIndex: 0 }));
   }, []);
 
-    // Handler for the switch filter changes
+  // Handler for the switch filter changes
   const handleShowOnlyMyRequirementsFiltersChange = useCallback(
     (filters: {
       checked: boolean;
@@ -386,7 +270,7 @@ function Requirements() {
       columnFilters?: MRT_TableState<InspectionRequirementGrid>["columnFilters"];
     }) => {
       setShowOnlyMyRequirements(filters.checked);
-      
+
       // Apply external filters
       Object.entries(filters.externalFilters).forEach(([key, value]) => {
         setExternalFilters((prev) => ({
@@ -394,148 +278,22 @@ function Requirements() {
           [key]: value,
         }));
       });
-      
+
       // Reset pagination
       setPagination((prev) => ({ ...prev, pageIndex: 0 }));
     },
     []
   );
 
-  const columns = useMemo<MRT_ColumnDef<InspectionRequirementGrid>[]>(
-    () => [
-      {
-        accessorFn: (row) => row.topic?.name,
-        id: "topic",
-        header: "Topic",
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          topics?.map((topic) => ({
-            text: topic.name,
-            value: topic.id.toString(),
-          })) ?? [],
-        size: 120,
-      },
-      {
-        accessorKey: "summary",
-        header: "Requirement Summary",
-        filterFn: "contains",
-        size: 200,
-      },
-      {
-        accessorFn: (row) => row.compliance_finding?.name,
-        id: "compliance_finding",
-        header: "Compliance Finding",
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          complianceFindings?.map((complianceFinding) => ({
-            text: complianceFinding.name,
-            value: complianceFinding.id.toString(),
-          })) ?? [],
-        size: 80,
-      },
-      {
-        accessorFn: (row) => row.enforcement_action?.name,
-        id: "enforcement_action",
-        header: "Enforcement Action",
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          enforcementActions?.map((enforcementAction) => ({
-            text: enforcementAction.name,
-            value: enforcementAction.id.toString(),
-          })) ?? [],
-        size: 150,
-      },
-      {
-        accessorKey: "approval_status",
-        header: "Approval Status",
-        Cell: ({ row }) => {
-          return row.original.approval_status ? (
-            <Chip
-              label={row.original.approval_status.name}
-              color={
-                row.original.approval_status.id ===
-                APPROVAL_STATUS.APPROVAL_PENDING
-                  ? "warning"
-                  : row.original.approval_status.id === APPROVAL_STATUS.APPROVED
-                    ? "success"
-                    : "error"
-              }
-              variant="outlined"
-              size="small"
-            />
-          ) : (
-            <></>
-          );
-        },
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          approvalStatusOptions?.map((approvalStatus) => ({
-            text: approvalStatus.name,
-            value: approvalStatus.id.toString(),
-          })) ?? [],
-        size: 120,
-      },
-      {
-        accessorFn: (row) => row.approved_by?.name,
-        id: "reviewer",
-        header: "Reviewer",
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          staffUsers?.map((staffUser) => ({
-            text: staffUser.name,
-            value: staffUser.id.toString(),
-          })) ?? [],
-        size: 100,
-      },
-      {
-        accessorKey: "requirement_number",
-        header: "Condition #",
-        filterFn: "contains",
-        size: 80,
-      },
-      {
-        accessorFn: (row) => row.requirement_source?.name,
-        id: "requirement_source",
-        header: "Source",
-        filterVariant: "multi-select",
-        filterSelectOptions:
-          requirementSources?.map((requirementSource) => ({
-            text: requirementSource.name,
-            value: requirementSource.id.toString(),
-          })) ?? [],
-        size: 100,
-      },
-      {
-        accessorKey: "ir_number",
-        header: "IR #",
-        filterFn: "contains",
-        Cell: ({ row }) => (
-          <PageLink
-            to="/ce-database/inspections/$inspectionNumber"
-            params={{ inspectionNumber: row.original.ir_number }}
-          />
-        ),
-        size: 120,
-      },
-      {
-        accessorFn: (row) =>
-          row.date_issued ? dateUtils.formatDate(row.date_issued) : "",
-        id: "date_issued",
-        header: "IR Issuance Date",
-        filterVariant: "date",
-        filterFn: "greaterThanOrEqual",
-        size: 120,
-      },
-    ],
-    [
-      topics,
-      complianceFindings,
-      enforcementActions,
-      requirementSources,
-      approvalStatusOptions,
-      staffUsers,
-    ]
-  );
+  // Use the extracted utility function for columns
+  const columns = useRequirementsGridColumns({
+    topics,
+    complianceFindings,
+    enforcementActions,
+    requirementSources,
+    approvalStatusOptions,
+    staffUsers,
+  });
 
   return isLoading || !isRestored ? (
     <Box
@@ -612,52 +370,11 @@ function Requirements() {
           >
             {/* Left side - Export button */}
             <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-              <Button
-                variant="text"
-                size="small"
-                startIcon={<FileDownloadRounded />}
-                sx={{ ml: -2 }}
-                onClick={() => downloadRequirementExport(queryParams)}
-              >
-                Export as Excel
-              </Button>
-              <Box
-                sx={{
-                  display: "flex",
-                  alignItems: "center",
-                  gap: 0.5,
-                  justifyContent: "center",
-                }}
-              >
-                <Typography variant="body1">
-                  {table.getState().pagination.pageIndex *
-                    table.getState().pagination.pageSize +
-                    1}{" "}
-                  to{" "}
-                  {Math.min(
-                    (table.getState().pagination.pageIndex + 1) *
-                      table.getState().pagination.pageSize,
-                    data?.total || 0
-                  )}{" "}
-                  of {data?.total || 0}
-                </Typography>
-                <IconButton
-                  aria-label="page_back"
-                  size="small"
-                  onClick={() => table.previousPage()}
-                  disabled={!table.getCanPreviousPage()}
-                >
-                  <ChevronLeftRounded fontSize="small" />
-                </IconButton>
-                <IconButton
-                  aria-label="page_forward"
-                  size="small"
-                  onClick={() => table.nextPage()}
-                  disabled={!table.getCanNextPage()}
-                >
-                  <ChevronRightRounded fontSize="small" />
-                </IconButton>
-              </Box>
+              <RequirementsGridExport queryParams={queryParams} />
+              <RequirementsGridPagination
+                table={table}
+                totalCount={data?.total || 0}
+              />
             </Box>
 
             {/* Right side - Filters */}
